move jid parameter down to negotiation impls
