+++ core/src/saros/negotiation/Negotiation.java
@@ -14,7 +14,6 @@ import saros.negotiation.NegotiationTools.CancelOption;
 import saros.net.IReceiver;
 import saros.net.ITransmitter;
 import saros.net.PacketCollector;
-import saros.net.xmpp.JID;
 
 /**
  * Abstract base class for implementing specific types of message exchanges within the Saros
@@ -36,9 +35,6 @@ abstract class Negotiation {
   private static final Logger log = Logger.getLogger(Negotiation.class);
 
   private final String id;
-
-  // FIMXE make this final (do not obtain the JID during the Negotiation !)
-  private JID peer;
 
   protected final ITransmitter transmitter;
 
@@ -70,10 +66,8 @@ abstract class Negotiation {
    * @param transmitter transmitter used for sending negotiation messages
    * @param receiver receiver used for receiving negotiation messages
    */
-  protected Negotiation(
-      String id, JID peer, final ITransmitter transmitter, final IReceiver receiver) {
+  protected Negotiation(String id, final ITransmitter transmitter, final IReceiver receiver) {
     this.id = id;
-    this.peer = peer;
     this.transmitter = transmitter;
     this.receiver = receiver;
   }
@@ -86,27 +80,6 @@ abstract class Negotiation {
    */
   public String getID() {
     return id;
-  }
-
-  /**
-   * Returns the JID of the peer with which the negotiation takes place.
-   *
-   * @return peer JID
-   */
-  public JID getPeer() {
-    return peer;
-  }
-
-  /**
-   * Changes the peer to negotiate with.
-   *
-   * @param peer new peer JID
-   * @deprecated The peer JID should remain constant during the negotiation. Code using this method
-   *     should be changed to find out the correct peer JID before the negotiation starts.
-   */
-  @Deprecated
-  protected void setPeer(JID peer) {
-    this.peer = peer;
   }
 
   /**
