+++ core/src/saros/negotiation/SessionNegotiation.java
@@ -66,6 +66,9 @@ public abstract class SessionNegotiation extends Negotiation {
 
   protected ISarosSession sarosSession;
 
+  // FIMXE make this final (do not obtain the JID during the Negotiation !)
+  private JID peer;
+
   public SessionNegotiation(
       final String id,
       final JID peer,
@@ -74,8 +77,9 @@ public abstract class SessionNegotiation extends Negotiation {
       final SessionNegotiationHookManager hookManager,
       final ITransmitter transmitter,
       final IReceiver receiver) {
-    super(id, peer, transmitter, receiver);
+    super(id, transmitter, receiver);
 
+    this.peer = peer;
     this.sessionManager = sessionManager;
     this.hookManager = hookManager;
     this.description = description;
@@ -107,5 +111,26 @@ public abstract class SessionNegotiation extends Negotiation {
   @Override
   protected void notifyTerminated(NegotiationListener listener) {
     listener.negotiationTerminated(this);
+  }
+
+  /**
+   * Returns the JID of the peer with which the negotiation takes place.
+   *
+   * @return peer JID
+   */
+  public JID getPeer() {
+    return peer;
+  }
+
+  /**
+   * Changes the peer to negotiate with.
+   *
+   * @param peer new peer JID
+   * @deprecated The peer JID should remain constant during the negotiation. Code using this method
+   *     should be changed to find out the correct peer JID before the negotiation starts.
+   */
+  @Deprecated
+  protected void setPeer(JID peer) {
+    this.peer = peer;
   }
 }
